moved "koch_snowflake.py" from "other" to "graphics"
